==> 02_geocoding_stats.ipynb <==
import pandas as pd
import numpy as np
import json, re

geodf_merged = pd.DataFrame.from_csv("output/geocoded_merged.csv")
market_df = pd.DataFrame.from_csv("output/no_geo.csv")

How many towns returned results?
================================

f_unique = len(geodf_merged.ix[geodf_merged.result_placename!='not_found'].hgr_id.unique())
nf_unique = len(geodf_merged.ix[geodf_merged.result_placename=='not_found'].hgr_id.unique())
total = len(geodf_merged.hgr_id.unique())
overlap = (nf_unique + f_unique) - total #overlap caused by changing spelling. 
print("{0} found\n{1} not found\n{2} total\n{3:.2%} returned results from geonames".format(f_unique,nf_unique-overlap,total,f_unique/total))

finding_counts = geodf_merged.ix[geodf_merged.result_placename!='not_found'].pivot_table(index='hgr_id',values='lat',aggfunc=lambda x: len(x.unique()))

fdf = pd.DataFrame(finding_counts)

print("{} found with one result".format(len(fdf.ix[fdf.lat==1])))
fdf.ix[fdf.lat!=1].describe()

Goldilocks stats
================

too_many = len(fdf.ix[fdf.lat!=1])
just_right = len(fdf.ix[fdf.lat==1])
no_results = len(geodf_merged.ix[geodf_merged.result_placename=='not_found'].hgr_id.unique())
total = len(market_df)
print("Too many (>1)\t{0}\t{1:.2%}\nJust right\t{2}\t{3:.2%}\nNo results from Geonames\t{4}\t{5:.2%}".format(too_many,too_many/total,just_right,just_right/total,no_results-overlap,(no_results-overlap)/total))

Precision by Namiestnichestvo
=============================

prov_precision_1 = pd.DataFrame(geodf_merged.ix[geodf_merged.result_placename != 'not_found'].pivot_table(index='prov_stem',values='hgr_id',aggfunc=lambda x: len(x.unique())))
prov_precision_2 = pd.DataFrame(geodf_merged.ix[geodf_merged.result_placename != 'not_found'].pivot_table(index='prov_stem',values='lat',aggfunc=lambda x: len(x.unique())))
prov_precision = pd.merge(prov_precision_1,prov_precision_2,left_index=True,right_index=True)
prov_precision['pct'] = prov_precision.lat/prov_precision.hgr_id*100
prov_precision.sort(columns='pct',ascending=False)

Recall by Namiestnichestvo
==========================

prov_recall_1 = pd.DataFrame(geodf_merged.pivot_table(index='prov_stem',values='hgr_id',aggfunc=lambda x: len(x.unique())))
prov_recall_2 = pd.DataFrame(geodf_merged.ix[geodf_merged.result_placename == 'not_found'].pivot_table(index='prov_stem',values='hgr_id',aggfunc=lambda x: len(x.unique())))
prov_recall = pd.merge(prov_recall_1,prov_recall_2,left_index=True,right_index=True)
prov_recall['pct'] = prov_recall.hgr_id_y/prov_recall.hgr_id_x*100
prov_recall.sort(columns='pct',ascending=False)
