==> 04_finalize_ortho_reverse_geocode.ipynb <==
Just the basics
===============
This notebook is intended to take a csv that doesn't have all of the information that we need in it, like transliterations, reverse geocoding, etc. and add it. It doesn't make any decisions, it just adds/edits information in an already pretty okay csv.

import time, requests, json, re
import pandas as pd
import numpy as np
from lxml import html

lookup_dict = {}
with open("special_cases.json",'r') as fp:
    lookup_dict = json.load(fp)

def ortho_rules(text,lookup_dict):
    """
    Takes a text string and a lookup dict as input
    Removes any lingering html tags
    Breaks the string down into component words with split
    Uses replacements to modernize orthography
    Returns re-assembled text with modern orthography
    """
    if text is not None:
        textlist = text.split(' ')
        out_list = []
        for word in textlist:
            if word in [',','.',';',':']:
                out_list.append(word)
            word = word.strip(',.;:')
            if word in lookup_dict:
                word = lookup_dict[word]
                out_list.append(word)
            else:
                word = re.sub("[Ъъ]$",'',word)
                word = re.sub(r"^([че]{0,2}[ч]?[вбнр]?[веоаи])з([ПФКТШСЧпфктшсч])",r"\1с\2",word,re.IGNORECASE)
                word = re.sub("аго$","ого",word,re.IGNORECASE)
                word = re.sub("яго$","его",word,re.IGNORECASE)
                word = re.sub("ыя$","ые",word,re.IGNORECASE)
                word = re.sub("iя$","ие",word,re.IGNORECASE)
                word = word.replace('i','и') #english letter
                word = word.replace('I','И') #english letter
                word = word.replace('і','и') #cyrillic letter
                word = word.replace('І','И') #cyrillic letter
                word = word.replace('Ѣ','Е')
                word = word.replace('ѣ','е')
                out_list.append(word)
        out_str = ' '.join(out_list)
        return out_str
    else:
        return "None"

cyrillic_translit={u'\u0410': 'A', u'\u0430': 'a',u'\u0411': 'B', u'\u0431': 'b',u'\u0412': 'V', u'\u0432': 'v',u'\u0413': 'G', u'\u0433': 'g',u'\u0414': 'D', u'\u0434': 'd',u'\u0415': 'E', u'\u0435': 'e',u'\u0416': 'Zh', u'\u0436': 'zh',u'\u0417': 'Z', u'\u0437': 'z',u'\u0418': 'I', u'\u0438': 'i',u'\u0419': 'I', u'\u0439': 'i',u'\u041a': 'K', u'\u043a': 'k',u'\u041b': 'L', u'\u043b': 'l',u'\u041c': 'M', u'\u043c': 'm',u'\u041d': 'N', u'\u043d': 'n',u'\u041e': 'O', u'\u043e': 'o',u'\u041f': 'P', u'\u043f': 'p',u'\u0420': 'R', u'\u0440': 'r',u'\u0421': 'S', u'\u0441': 's',u'\u0422': 'T', u'\u0442': 't',u'\u0423': 'U', u'\u0443': 'u',u'\u0424': 'F', u'\u0444': 'f',u'\u0425': 'Kh', u'\u0445': 'kh',u'\u0426': 'Ts', u'\u0446': 'ts',u'\u0427': 'Ch', u'\u0447': 'ch',u'\u0428': 'Sh', u'\u0448': 'sh',u'\u0429': 'Shch', u'\u0449': 'shch',u'\u042a': '"', u'\u044a': '"',u'\u042b': 'Y', u'\u044b': 'y',u'\u042c': "'", u'\u044c': "'",u'\u042d': 'E', u'\u044d': 'e',u'\u042e': 'Iu', u'\u044e': 'iu',u'\u042f': 'Ia', u'\u044f': 'ia',u'\u0462': 'E', u'\u0463': 'e'}

def transliterate(word, translit_table):
    """
    Transliterates 'word' based on the key/value pairs in 'translit_table'
    """
    converted_word = ''
    for char in word:
        transchar = ''
        if char in translit_table:
            transchar = translit_table[char]
        else:
            transchar = char
        converted_word += transchar
    return converted_word

def reverse_geocode(lng,lat):
    keepgoing = True
    while keepgoing == True:
        try:
            url = "https://maps.googleapis.com/maps/api/geocode/json"
            Q = {'latlng':"{},{}".format(lat,lng)}
            R = requests.get(url, params=Q)
            print(R.url)
            geo_response = R.json()
            locality = 'error'
            for component in geo_response['results'][0]['address_components']:
                if 'locality' in component['types']:
                    locality = component['long_name']
                elif 'administrative_area_level_1' in component['types']:
                    admin1 = component['long_name']
                elif 'country' in component['types']:
                    country = component['long_name']
                    countryCode = component['short_name']
            if locality == 'error':
                for component in geo_response['results'][0]['address_components']:
                    if 'administrative_area_level_2' in component['types']:
                        locality = component['long_name']
            print("\"{0}, {1}\",\"{2}\",\"{3}\"".format(locality,admin1,country,countryCode))
            time.sleep(1)
            keepgoing = False
        except IndexError:
            time.sleep(1)
            continue
        return "{}, {}".format(locality,admin1), country, countryCode

finalizer = pd.DataFrame.from_csv("hgr_test_set.csv",index_col=None)
finalizer.head()

finalizer['ru_new_orth'] = finalizer.ru_old_orth.apply(ortho_rules,args=[lookup_dict])

finalizer['lc_trans'] = finalizer.ru_new_orth.apply(transliterate,args=[cyrillic_translit])

finalizer['pres_loc'], finalizer['country'], finalizer['country_code'] = np.vectorize(reverse_geocode)(finalizer.x_coord,finalizer.y_coord)

finalizer.to_csv("finalized.csv",encoding="utf-8")

This code block preserved because it's the original reverse geocoding block. Not for production use.
====================================================================================================

namiest = [[46.5813,54.8421],[37.0674,54.50503],[43.83992,55.39485],[34.89879,50.31036],[30.5725,54.8351],[36.69651,57.78506],[36.13194,53.81194],[30.4122,52.9395],[38.51663,54.37445],[37.8084,60.0288],[38.1067,50.2035],[30.54517,56.34],[38.26843,54.35533],[36.17194,55.34472],[107.6,51.833333],[39.88398,59.2239],[35.96944,56.03361],[28.75283,60.70763],[33.99099,50.37112],[42.69351,56.20152],[26.53333,55.88333],[37.52041,56.34485],[39.15553,53.25155],[34.98333,48.45],[38.50169,52.62366],[38.55181,53.82526],[38.87841,54.7605],[36.85918,55.73402],[34.58935,56.17534],[37.24848,49.20875],[37.24849,49.20875],[41.39528,54.94111],[37.60806,57.35917],[38.16694,54.84444],[43.05744,53.69157],[42.12894,57.43914],[108.11194,57.78528],[30.5238,50.45466],[31.9586,53.6079],[35.77159,54.03746],[40.4907,52.8978],[38.77833,55.07944],[37.18598,50.8132],[40.92686,57.76647],[43.78963,54.4253],[36.18735,51.73733],[39.16639,53.01942],[37.60689,52.42534],[32.99694,50.01625],[40.68433,58.3625],[44.2542,65.8401],[33.60886,49.96854],[35.26233,51.023],[36.02722,55.50194],[31.7217,54.0185],[43.69458,53.87654],[43.67663,53.53041],[31.27104,58.52131],[31.27104,58.52131],[33.266667,51.983333],[37.04048,52.97387],[31.88688,51.04801],[31.88688,51.04801],[30.4172,54.5081],[28.35368,57.34395],[31.4795,50.08768],[39.26928,57.01039],[32.50621,50.24242],[34.54073,49.59373],[28.7856,55.4879],[39.13531,58.49928],[32.38761,50.59323],[28.3496,57.8136],[30.0495,53.0934],[39.41394,57.1914],[36.19522,55.69898],[44.7329,53.1783],[45.1749,54.1838],[46.00861,51.54056],[48.66051,55.77329],[48.38657,54.32824],[32.49985,51.52387],[30.2454,53.521],[33.67942,52.2405],[40.44668,56.42274],[35.90057,56.85836],[43.21483,54.63023],[37.616667,54.2],[36.25272,49.98081],[61.42915,55.15402],[31.3831,53.5689],[31.28488,51.50551],[42.68318,58.75375],[41.71191,54.02576],[43.1041,57.32007],[49.09162,62.16755],[39.87368,57.62987]]
namiest_locations = []
for x in namiest:
    keepgoing = True
    while keepgoing == True:
        try:
            url = "https://maps.googleapis.com/maps/api/geocode/json"
            Q = {'latlng':"{},{}".format(x[1],x[0])}
            R = requests.get(url, params=Q)
            #print(R.url)
            geo_response = R.json()
            locality = 'error'
            for component in geo_response['results'][0]['address_components']:
                if 'locality' in component['types']:
                    locality = component['long_name']
                elif 'administrative_area_level_1' in component['types']:
                    admin1 = component['long_name']
                elif 'country' in component['types']:
                    country = component['long_name']
                    countryCode = component['short_name']
            if locality == 'error':
                for component in geo_response['results'][0]['address_components']:
                    if 'administrative_area_level_2' in component['types']:
                        locality = component['long_name']
            outlist = [x[0],x[1],"{}, {}".format(locality,admin1),country,countryCode]
            namiest_locations.extend(outlist)
            print("\"{0}, {1}\",\"{2}\",\"{3}\"".format(locality,admin1,country,countryCode))
            time.sleep(1)
            keepgoing = False
        except IndexError:
            continue
