==> 05_geojson_map.ipynb <==
import pandas as pd
import numpy as np
import json

test_set = pd.DataFrame.from_csv("output/hgr_test_set_5.csv")
mydict = test_set.partof_id.to_dict()

geojson = {"type":"FeatureCollection","features":[]}

test_set.ix[np.isnan(test_set.partof_prov)]

prov_types = test_set[test_set.partof_prov == test_set.index].ru_featuretype.unique().tolist()
uezd_types = test_set[test_set.partof_uezd == test_set.index].ru_featuretype.unique().tolist()
print(uezd_types)

def mapmaker(name,desc,url,lat,lng,featuretype,partof_namest):
    if featuretype in ['Губернія', 'главный городъ', 'Намѣстничество', 'Область']:
        markerSize = 'large'
    elif featuretype in ['уезд', 'городъ', 'уезде города']:
        markerSize = 'medium'
    else:
        markerSize = 'small'
    if featuretype in ['Губернія', 'главный городъ', 'Намѣстничество', 'Область']:
        markerSymbol = 'star'
    elif featuretype in ['уезд', 'городъ', 'уезде города']:
        markerSymbol = 'square'
    else:
        markerSymbol = 'circle'
    colorlist = ["#D29176","#62E53A","#D353D9","#6EDCCF","#522A5A","#DACA3C","#394D1C","#DD3E27","#6B8FDB","#CE3568","#48972D","#D2C7C4","#E69631","#405A6D","#8C2D29","#6BE29C","#D0A2D2","#BCDC7C","#A359A8","#90936A","#81BAD9","#674424","#302C2B","#7F68D6","#9C782B","#BDDCB1","#4F9D66","#BA602C","#907178","#5F2333","#CB6E8A","#B2DD3B","#555B92","#DBC277","#DF5B59","#609797","#7B8B2E","#CB4094","#37624B","#6BD960"]
    if int(partof_namest) == 263:
        partof_namest = 40
    markerColor = markerColor = colorlist[int(partof_namest-1)]
    if str(desc)=='nan':
        desc = ""
    if str(url)=='nan':
        url = ""
    outdict = {"type":"Feature","geometry":{"type":"Point","coordinates":[lng,lat]},"properties":{"title":name,"description":desc+" "+url,"marker-size":markerSize,"marker-symbol":markerSymbol,"marker-color":markerColor}}
    return outdict

geojson['features']=np.vectorize(mapmaker)(test_set.ru_old_orth,test_set.description,test_set.SOURCE,test_set.y_coord,test_set.x_coord,test_set.ru_featuretype,test_set.partof_prov).tolist()

with open("output/hgr_map_test.geojson","w") as fp:
    json.dump(geojson,fp)
